--- js/components/HomeScene/Scene/Rig/index.jsx
@@ -1,1 +1,0 @@
-export { default } from "./Rig";
